requirements.txt
gradio
fitz
Pillow
google-generativeai
python-dotenv
PyMuPDF